Introduce 2D Chart Data Model abstraction
